Cleaned up the view story view. This code is still pretty ugly / cluttered. Need to find a way to break them out into elements or something.
Completed pagination on entries on the view story page. (There was no open issue for this, but there should have been...) At some point we may want to have page:last links.

Fixed moderator routing. Apparently named parameters completely break reverse routing when using prefix routing in cakephp 1.2 unless you explicitly identify which named params are- allowed. I feel like there's a better fix for this, but I don't have the time to hunt it down.

Fixed a previously unknown bug that created duplicate characters entries when a character was deactivated.

The bummer of this commit is it resolves no issues. It only takes care of some todos documented in the html...