Fixed a few slugging issues; there are still more.
Implemented a rudimentary turn cycling, that needs work. I hate the way this is being done; I need to plan this out more. Also, I need to send out messages to members of the story when an entry is posted.

Implemented some basic controls to limit a user from adding a character to multiple stories. The backend code is a bit gross, but it works.

Also removed two unused actions (whoops).

Closes #86
Closes #83
Closes #88