As part of working on the messaging system, I moved it to its own component where I could write a test case for it. It now is automatically being tested in a very low level form, but should any bugs turn up, I'll be able to avoid a regression. This is particularly important, because I don't want to be spamming users.
During that change, I implemented the riiga parser using a bit of markdown sytax and some custom tagging / ooc comments bracketing. Testing is in place for that helper as well, which is particularly nice, because I found a regression due to an off by one in the string parsing. I suspect more issues will turn up eventually. I went ahead and implemented that everywhere. One addition to the entry controller was a redirector to a particular entry in a story. This meant adding anchors on the view story page that would pin an entry.

To take advantage of the parser, emails now support both html and plain text (so does the parser).

And as part of that, added a view entry that redirects to the correct story page with an anchor to the entry.

Altered the themeing a little to make it more readable in a few areas.

Added basic date helper to centralize date logic. Added testing for this as well.